De-namespace schema types
